fixed reversed depth data
