+++ src/main.rs
@@ -22,6 +22,7 @@ pub mod moves;
 pub mod piece_sq_tables;
 pub mod types;
 pub mod game;
+pub mod uci;
 
 #[global_allocator]
 static GLOBAL: MiMalloc = MiMalloc;
@@ -133,7 +134,8 @@ fn output_thread(out: UciMessage, toggle_ready_ok: &Arc<RwLock<bool>>) {
                     let mut bot = BasicBot::new(&board);
                     let (eval, chess_move) = bot.search(depth as u16);
                     let best_uci_move = chess_move_to_uci_move(&chess_move);
-                    let depth_data = bot.get_depth_data();
+
+                    let depth_data = bot.uci.get_depth_data();
 
                     for data in depth_data {
                         let mut info_vec: Vec<UciInfoAttribute> = vec![];
